requirements.txt
python-telegram-bot==20.7
requests==2.32.3
Flask==3.0.3